some minor changes: - do not disable replaygain calc and find peak sample, when not writing LAME info header (patch tracker item #66) - remove deprecated linker option from Makefile.MSVC (bug tracker item #442)

+++ frontend/get_audio.c
@@ -1335,7 +1335,7 @@ static short const WAVE_FORMAT_PCM = 0x0001;
 static short const WAVE_FORMAT_IEEE_FLOAT = 0x0003;
 #endif
 #ifndef WAVE_FORMAT_EXTENSIBLE
-static short const WAVE_FORMAT_EXTENSIBLE = (short)0xFFFE;
+static short const WAVE_FORMAT_EXTENSIBLE = 0xFFFE;
 #endif
